Better Semaphore Lease Contention Handling
Add the same retry logic from `AcquireSemaphore` to `CancelSemaphoreLease`
to handle contetion. Without the retry it is possible for a cancellation
to fail and the lease in question to remain held for its entire expiry.
If the number of cancellations that fail is >= `max_connections` then
this causes a user to effectively be locked out until the leases are
expired.

Fixes #10363